+++ src/CommunityToolkit.Maui/CommunityToolkit.Maui.csproj
@@ -55,6 +55,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <None Remove="Essentials\" />
+    <None Remove="Essentials\AppTheme\" />
+  </ItemGroup>
+  <ItemGroup>
     <None Include="..\CommunityToolkit.Maui.SourceGenerators\bin\$(Configuration)\netstandard2.0\CommunityToolkit.Maui.SourceGenerators.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
     <None Include="..\CommunityToolkit.Maui.Analyzers\bin\$(Configuration)\netstandard2.0\CommunityToolkit.Maui.Analyzers.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
     <None Include="..\CommunityToolkit.Maui.Analyzers.CodeFixes\bin\$(Configuration)\netstandard2.0\CommunityToolkit.Maui.Analyzers.CodeFixes.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
@@ -72,4 +76,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Essentials\" />
+    <Folder Include="Essentials\AppTheme\" />
+  </ItemGroup>
 </Project>
