+++ src/CommunityToolkit.Maui/Essentials/AppTheme/AppThemeResource.shared.cs
